chore(controllers): move TestLabelMatchedExpression to another file
