+++ controllers/controller_shared_test.go
@@ -1,0 +1,175 @@
+/*
+Copyright 2022.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controllers
+
+import (
+	"testing"
+
+	"github.com/grafana/grafana-operator/v5/api/v1beta1"
+	"github.com/stretchr/testify/assert"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func TestLabelMatchedExpression(t *testing.T) {
+	dashboardList := v1beta1.GrafanaDashboardList{
+		TypeMeta: metav1.TypeMeta{},
+		ListMeta: metav1.ListMeta{},
+		Items: []v1beta1.GrafanaDashboard{
+			{
+				TypeMeta: metav1.TypeMeta{},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "grafana-match-labels",
+					Namespace: "grafana-operator-system",
+				},
+				Spec: v1beta1.GrafanaDashboardSpec{
+					InstanceSelector: &metav1.LabelSelector{
+						MatchLabels: map[string]string{
+							"dashboard": "grafana",
+						},
+					},
+				},
+			},
+			{
+				TypeMeta: metav1.TypeMeta{},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "grafana-exists",
+					Namespace: "grafana-operator-system",
+				},
+				Spec: v1beta1.GrafanaDashboardSpec{
+					InstanceSelector: &metav1.LabelSelector{
+						MatchExpressions: []metav1.LabelSelectorRequirement{
+							{
+								Operator: metav1.LabelSelectorOpExists,
+								Key:      "dashboard",
+							},
+						},
+					},
+				},
+			},
+			{
+				TypeMeta: metav1.TypeMeta{},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "grafana-not-exists",
+					Namespace: "grafana-operator-system",
+				},
+				Spec: v1beta1.GrafanaDashboardSpec{
+					InstanceSelector: &metav1.LabelSelector{
+						MatchExpressions: []metav1.LabelSelectorRequirement{
+							{
+								Operator: metav1.LabelSelectorOpDoesNotExist,
+								Key:      "dashboard",
+							},
+						},
+					},
+				},
+			},
+			{
+				TypeMeta: metav1.TypeMeta{},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "grafana-not-in",
+					Namespace: "grafana-operator-system",
+				},
+				Spec: v1beta1.GrafanaDashboardSpec{
+					InstanceSelector: &metav1.LabelSelector{
+						MatchExpressions: []metav1.LabelSelectorRequirement{
+							{
+								Operator: metav1.LabelSelectorOpNotIn,
+								Key:      "dashboard",
+								Values: []string{
+									"grafana-1",
+								},
+							},
+						},
+					},
+				},
+			},
+			{
+				TypeMeta: metav1.TypeMeta{},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "grafana-in",
+					Namespace: "grafana-operator-system",
+				},
+				Spec: v1beta1.GrafanaDashboardSpec{
+					InstanceSelector: &metav1.LabelSelector{
+						MatchExpressions: []metav1.LabelSelectorRequirement{
+							{
+								Operator: metav1.LabelSelectorOpIn,
+								Key:      "dashboard",
+								Values: []string{
+									"grafana-1",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+	grafanaList := []v1beta1.Grafana{
+		{
+			TypeMeta: metav1.TypeMeta{},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "grafana",
+				Namespace: "grafana-operator-system",
+				Labels: map[string]string{
+					"dashboard": "grafana",
+				},
+			},
+			Status: v1beta1.GrafanaStatus{
+				Dashboards: v1beta1.NamespacedResourceList{
+					"grafana-match-labels/matchLabels",
+					"grafana-not-in/notin",
+					"grafana-exists/exists",
+				},
+			},
+		},
+		{
+			TypeMeta: metav1.TypeMeta{},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "grafana-1",
+				Namespace: "grafana-operator-system",
+				Labels: map[string]string{
+					"dashboard": "grafana-1",
+				},
+			},
+			Status: v1beta1.GrafanaStatus{
+				Dashboards: v1beta1.NamespacedResourceList{
+					"grafana-in/in",
+					"grafana-exists/exists",
+				},
+			},
+		},
+	}
+
+	for _, dashboard := range dashboardList.Items {
+		var selectedList v1beta1.GrafanaList
+		for _, instance := range grafanaList {
+			selected := labelMatchedExpression(instance, dashboard.Spec.InstanceSelector)
+			if selected {
+				selectedList.Items = append(selectedList.Items, instance)
+			}
+		}
+		for _, instance := range selectedList.Items {
+			if instance.Name == "grafana" {
+				assert.Equal(t, v1beta1.NamespacedResourceList{"grafana-match-labels/matchLabels", "grafana-not-in/notin", "grafana-exists/exists"}, instance.Status.Dashboards)
+			}
+			if instance.Name == "grafana-1" {
+				assert.Equal(t, v1beta1.NamespacedResourceList{"grafana-in/in", "grafana-exists/exists"}, instance.Status.Dashboards)
+			}
+		}
+	}
+}
