+++ controllers/dashboard_controller_test.go
@@ -155,153 +155,3 @@ func TestGetDashboardEnvs(t *testing.T) {
 	assert.NotNil(t, envs)
 	assert.True(t, len(envs) == 1, "Expected 1 env, got %d", len(envs))
 }
-
-func TestLabelMatchedExpression(t *testing.T) {
-	dashboardList := v1beta1.GrafanaDashboardList{
-		TypeMeta: metav1.TypeMeta{},
-		ListMeta: metav1.ListMeta{},
-		Items: []v1beta1.GrafanaDashboard{
-			{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "grafana-match-labels",
-					Namespace: "grafana-operator-system",
-				},
-				Spec: v1beta1.GrafanaDashboardSpec{
-					InstanceSelector: &metav1.LabelSelector{
-						MatchLabels: map[string]string{
-							"dashboard": "grafana",
-						},
-					},
-				},
-			},
-			{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "grafana-exists",
-					Namespace: "grafana-operator-system",
-				},
-				Spec: v1beta1.GrafanaDashboardSpec{
-					InstanceSelector: &metav1.LabelSelector{
-						MatchExpressions: []metav1.LabelSelectorRequirement{
-							{
-								Operator: metav1.LabelSelectorOpExists,
-								Key:      "dashboard",
-							},
-						},
-					},
-				},
-			},
-			{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "grafana-not-exists",
-					Namespace: "grafana-operator-system",
-				},
-				Spec: v1beta1.GrafanaDashboardSpec{
-					InstanceSelector: &metav1.LabelSelector{
-						MatchExpressions: []metav1.LabelSelectorRequirement{
-							{
-								Operator: metav1.LabelSelectorOpDoesNotExist,
-								Key:      "dashboard",
-							},
-						},
-					},
-				},
-			},
-			{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "grafana-not-in",
-					Namespace: "grafana-operator-system",
-				},
-				Spec: v1beta1.GrafanaDashboardSpec{
-					InstanceSelector: &metav1.LabelSelector{
-						MatchExpressions: []metav1.LabelSelectorRequirement{
-							{
-								Operator: metav1.LabelSelectorOpNotIn,
-								Key:      "dashboard",
-								Values: []string{
-									"grafana-1",
-								},
-							},
-						},
-					},
-				},
-			},
-			{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "grafana-in",
-					Namespace: "grafana-operator-system",
-				},
-				Spec: v1beta1.GrafanaDashboardSpec{
-					InstanceSelector: &metav1.LabelSelector{
-						MatchExpressions: []metav1.LabelSelectorRequirement{
-							{
-								Operator: metav1.LabelSelectorOpIn,
-								Key:      "dashboard",
-								Values: []string{
-									"grafana-1",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-	grafanaList := []v1beta1.Grafana{
-		{
-			TypeMeta: metav1.TypeMeta{},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "grafana",
-				Namespace: "grafana-operator-system",
-				Labels: map[string]string{
-					"dashboard": "grafana",
-				},
-			},
-			Status: v1beta1.GrafanaStatus{
-				Dashboards: v1beta1.NamespacedResourceList{
-					"grafana-match-labels/matchLabels",
-					"grafana-not-in/notin",
-					"grafana-exists/exists",
-				},
-			},
-		},
-		{
-			TypeMeta: metav1.TypeMeta{},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "grafana-1",
-				Namespace: "grafana-operator-system",
-				Labels: map[string]string{
-					"dashboard": "grafana-1",
-				},
-			},
-			Status: v1beta1.GrafanaStatus{
-				Dashboards: v1beta1.NamespacedResourceList{
-					"grafana-in/in",
-					"grafana-exists/exists",
-				},
-			},
-		},
-	}
-
-	for _, dashboard := range dashboardList.Items {
-		var selectedList v1beta1.GrafanaList
-		for _, instance := range grafanaList {
-			selected := labelMatchedExpression(instance, dashboard.Spec.InstanceSelector)
-			if selected {
-				selectedList.Items = append(selectedList.Items, instance)
-			}
-		}
-		for _, instance := range selectedList.Items {
-			if instance.Name == "grafana" {
-				assert.Equal(t, v1beta1.NamespacedResourceList{"grafana-match-labels/matchLabels", "grafana-not-in/notin", "grafana-exists/exists"}, instance.Status.Dashboards)
-			}
-			if instance.Name == "grafana-1" {
-				assert.Equal(t, v1beta1.NamespacedResourceList{"grafana-in/in", "grafana-exists/exists"}, instance.Status.Dashboards)
-			}
-		}
-	}
-}
